Perform some code cleanup
